refactor: 변수명 수정
