+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/memberLecture/MemoRequest.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.dto.lecture;
+package com.ktnu.AiLectureSummary.dto.memberLecture;
 
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class PersonalNoteRequest {
+public class MemoRequest {
     @Size(max = 1000, message = "메모는 최대 1000자 까지 입력 가능합니다.")
-    private String note;
+    private String memo;
 }
